requirements.txt
amqp==1.4.9
anyjson==0.3.3
atomicwrites==1.3.0
attrs==19.1.0
Babel==2.6.0
billiard==3.3.0.23
celery==3.1.26.post2
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.1
django-celery==3.2.2
django-cors-headers==2.5.3
django-extensions==2.1.6
django-filter==2.1.0
django-js-asset==1.2.2
django-mptt==0.10.0
django-phonenumber-field==2.3.1
django-pytest==0.2.0
djangorestframework==3.9.3
djangorestframework-filters==0.11.1
drf-yasg==1.15.0
factory-boy==2.11.1
Faker==1.0.5
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
kombu==3.0.37
MarkupSafe==1.1.1
more-itertools==7.0.0
phonenumbers==8.10.10
pluggy==0.9.0
psycopg2-binary==2.8.2
py==1.8.0
pydot==1.4.1
pyparsing==2.4.0
pytest==4.4.1
pytest-django==3.4.8
pytest-factoryboy==2.0.2
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
ruamel.yaml==0.15.94
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
uritemplate==3.0.0
urllib3==1.24.3
